% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_std_sf_standardization.R
\name{convert.LL}
\alias{convert.LL}
\title{Alias for \code{\link{convert_lengths}}}
\usage{
convert.LL(
  species_code,
  length_from_code,
  length_to_code = "FL",
  measurement,
  ll_equations = DEFAULT_IOTC_LL_EQUATIONS
)
}
\description{
Alias for \code{\link{convert_lengths}}
}
