% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_std_sf_standardization.R
\name{convert.LW}
\alias{convert.LW}
\title{Alias for \code{\link{length_to_weight}}}
\usage{
convert.LW(
  species_code,
  gear_type = GEAR_TYPE_PSPLGI,
  length_from_code = "FL",
  weight_to_code = "RND",
  measurement,
  lw_equations = DEFAULT_IOTC_LW_EQUATIONS
)
}
\description{
Alias for \code{\link{length_to_weight}}
}
