% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_std_sf_standardization.R
\name{convert.WL.ND}
\alias{convert.WL.ND}
\title{Alias for \code{\link{weight_to_length_non_deterministic}}}
\usage{
convert.WL.ND(
  species_code,
  weight_from_code = "RND",
  length_to_code = "FL",
  measurement,
  wl_keys = DEFAULT_IOTC_WL_ND_KEYS
)
}
\description{
Alias for \code{\link{weight_to_length_non_deterministic}}
}
