% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_std_sf_standardization.R
\name{convert.WL}
\alias{convert.WL}
\title{Alias for \code{\link{weight_to_length}}}
\usage{
convert.WL(
  species_code,
  gear_type = GEAR_TYPE_PSPLGI,
  weight_from_code = "RND",
  length_to_code = "FL",
  measurement,
  wl_equations = DEFAULT_IOTC_WL_EQUATIONS
)
}
\description{
Alias for \code{\link{weight_to_length}}
}
