% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_std_sf_standardization.R
\name{convert_lengths}
\alias{convert_lengths}
\title{Converts between length measurements for a given species (assuming an equation is known)}
\usage{
convert_lengths(
  species_code,
  length_from_code,
  length_to_code = "FL",
  measurement,
  ll_equations = DEFAULT_IOTC_LL_EQUATIONS
)
}
\description{
Converts between length measurements for a given species (assuming an equation is known)
}
