% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEFAULT_IOTC_LL_EQUATIONS}
\alias{DEFAULT_IOTC_LL_EQUATIONS}
\title{DEFAULT_IOTC_LL_EQUATIONS}
\format{
A data table
\describe{
  \item{\code{SPECIES}}{character The IOTC species code}
  \item{\code{FROM}}{character An IOTC length measure code}
  \item{\code{TO}}{character The IOTC length measure code for the species' standard length (one among: FL, EFL)}
  \item{\code{C}}{double the multiplicative constant for the conversion equation}
  \item{\code{A}}{double the 'A' costant for the conversion equation}
  \item{\code{B}}{double the 'B' costant for the conversion equation}
  \item{\code{EQ_ID}}{character the identifier for the conversion equation to use (one among: IDENTITY, SQUARED, PROP, INVPROP, INVPROP_ALT, POW, INVPOW)}
  \item{\code{NOTES}}{character Additional notes}
}
}
\source{
\url{https://iotc.org/WPTT/24/Data/13-Equations/}
}
\usage{
DEFAULT_IOTC_LL_EQUATIONS
}
\description{
The length-length conversion equations for the most common IOTC species
}
\keyword{datasets}
