% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_std_sf_standardization.R
\docType{data}
\name{GEAR_TYPE_LLOT}
\alias{GEAR_TYPE_LLOT}
\title{String constant for longline and all other gears}
\format{
An object of class \code{character} of length 1.
}
\usage{
GEAR_TYPE_LLOT
}
\description{
String constant for longline and all other gears
}
\keyword{datasets}
