% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_std_sf_standardization.R
\docType{data}
\name{GEAR_TYPE_PSPLGI}
\alias{GEAR_TYPE_PSPLGI}
\title{String constant for purse-seine, pole-and-line and gillnet gears}
\format{
An object of class \code{character} of length 1.
}
\usage{
GEAR_TYPE_PSPLGI
}
\description{
String constant for purse-seine, pole-and-line and gillnet gears
}
\keyword{datasets}
