% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_std_sf_standardization.R
\name{length_to_weight}
\alias{length_to_weight}
\title{Converts between length and weight measurements for a given species (assuming an equation is known)}
\usage{
length_to_weight(
  species_code,
  gear_type = GEAR_TYPE_PSPLGI,
  length_from_code = "FL",
  weight_to_code = "RND",
  measurement,
  lw_equations = DEFAULT_IOTC_LW_EQUATIONS
)
}
\description{
Converts between length and weight measurements for a given species (assuming an equation is known)
}
