% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_std_sf_standardization.R
\name{standardize.SF}
\alias{standardize.SF}
\title{Alias for \code{\link{standardize_size_frequencies}}}
\usage{
standardize.SF(
  raw_data,
  max_bin_size = 5,
  first_class_low = 10,
  last_size_bin = 150,
  bin_size = 2,
  keep_sex_and_raise_code = FALSE,
  ll_equations = DEFAULT_IOTC_LL_EQUATIONS,
  lw_equations = DEFAULT_IOTC_LW_EQUATIONS,
  wl_equations = DEFAULT_IOTC_WL_EQUATIONS,
  wl_keys = DEFAULT_IOTC_WL_ND_KEYS
)
}
\description{
Alias for \code{\link{standardize_size_frequencies}}
}
