% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_std_sf_standardization.R
\name{standardize_and_pivot_size_frequencies}
\alias{standardize_and_pivot_size_frequencies}
\title{Standardizes 'raw' SF data (in _wide_ format)}
\usage{
standardize_and_pivot_size_frequencies(
  raw_data,
  max_bin_size = 5,
  first_class_low = 10,
  last_size_bin = 150,
  bin_size = 2,
  keep_sex_and_raise_code = FALSE,
  ll_equations = DEFAULT_IOTC_LL_EQUATIONS,
  lw_equations = DEFAULT_IOTC_LW_EQUATIONS,
  wl_equations = DEFAULT_IOTC_WL_EQUATIONS,
  wl_keys = DEFAULT_IOTC_WL_ND_KEYS
)
}
\arguments{
\item{raw_data}{A single-species subset of the available SF data}

\item{max_bin_size}{The maximum allowed bin size (in cm) for the original data}

\item{first_class_low}{The maximum length (in cm) aggregated under the first size bin}

\item{last_size_bin}{The last size bin in the output}

\item{bin_size}{The size (in cm) of the output size bins}

\item{keep_sex_and_raise_code}{Whether to stratify or not the output by sex and raise code (defaults to \code{FALSE})}

\item{ll_equations}{The deterministic L-L conversion equations. Defaults to \code{\link{DEFAULT_IOTC_LL_EQUATIONS}}}

\item{lw_equations}{The deterministic L-W conversion equations. Defaults to \code{\link{DEFAULT_IOTC_LW_EQUATIONS}}}

\item{wl_equations}{The deterministic W-L conversion equations. Defaults to \code{\link{DEFAULT_IOTC_WL_EQUATIONS}}}

\item{wl_keys}{The non-deterministic W-L keys.             Defaults to \code{\link{DEFAULT_IOTC_WL_ND_KEYS}}}
}
\value{
the resulting standardized dataset (in _wide_ format)
}
\description{
Input records undergo a number of processing steps that yield a SF dataset with all
records converted to a standard length measurement (species-specific), assigned to a size bin and complemented by
a calculated round weight (if proper L-W equations exist).
}
\details{
The conversion process consists in:

1) Removing all records with a non-standard (i.e., larger than \code{max_bin_size}) bin size

2) Converting all length measurements to the species-specific standard lengths (through \code{ll_equations})

3) Converting (deterministically) all weight measurements to species-specific standard lengths (through \code{wl_equations})

4) Converting (non-deterministically) all weight measurements to species-specific standard lengths (through \code{wl_keys})

5) Calculating the round weight for all records (through \code{lw_equations})

6) Assigning all records to a size bin of \code{bin_size} cm, collapsing all records with a length less than or equal to \code{first_class_low} cm to the first bin, and capping larger size bins to \code{last_size_bin}

7) Pivoting the results by size bin (from \code{T001} to \code{T<last_size_bin>})
}
