% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_std_sf_standardization.R
\name{standardize_size_frequencies}
\alias{standardize_size_frequencies}
\title{DESCRIPTION TO BE ADDED}
\usage{
standardize_size_frequencies(
  raw_data,
  max_bin_size = 5,
  first_class_low = 10,
  last_size_bin = 150,
  bin_size = 2,
  keep_sex_and_raise_code = FALSE,
  ll_equations = DEFAULT_IOTC_LL_EQUATIONS,
  lw_equations = DEFAULT_IOTC_LW_EQUATIONS,
  wl_equations = DEFAULT_IOTC_WL_EQUATIONS,
  wl_keys = DEFAULT_IOTC_WL_ND_KEYS
)
}
\arguments{
\item{raw_data}{A single-species subset of the available SF data}

\item{max_bin_size}{The maximum allowed bin size (in cm) for the original data}

\item{first_class_low}{The maximum length (in cm) aggregated under the first size bin}

\item{last_size_bin}{The last size bin in the output}

\item{bin_size}{The size (in cm) of the output size bins}

\item{keep_sex_and_raise_code}{Whether to stratify or not the output by sex and raise code (defaults to \code{FALSE})}

\item{ll_equations}{The deterministic L-L conversion equations. Defaults to \code{\link{DEFAULT_IOTC_LL_EQUATIONS}}}

\item{lw_equations}{The deterministic L-W conversion equations. Defaults to \code{\link{DEFAULT_IOTC_LW_EQUATIONS}}}

\item{wl_equations}{The deterministic W-L conversion equations. Defaults to \code{\link{DEFAULT_IOTC_WL_EQUATIONS}}}

\item{wl_keys}{The non-deterministic W-L keys.             Defaults to \code{\link{DEFAULT_IOTC_WL_ND_KEYS}}}
}
\description{
DESCRIPTION TO BE ADDED
}
