% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_std_sf_standardization.R
\name{weight_to_length}
\alias{weight_to_length}
\title{Converts between weight and length measurements for a given species (assuming an equation is known)}
\usage{
weight_to_length(
  species_code,
  gear_type = GEAR_TYPE_PSPLGI,
  weight_from_code = "RND",
  length_to_code = "FL",
  measurement,
  wl_equations = DEFAULT_IOTC_WL_EQUATIONS
)
}
\description{
Converts between weight and length measurements for a given species (assuming an equation is known)
}
