% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_base_common_std_sf_standardization.R
\name{weight_to_length_non_deterministic}
\alias{weight_to_length_non_deterministic}
\title{Converts between weight and length measurements for a given species (using non-deterministic conversion keys, assuming these are known)}
\usage{
weight_to_length_non_deterministic(
  species_code,
  weight_from_code = "RND",
  length_to_code = "FL",
  measurement,
  wl_keys = DEFAULT_IOTC_WL_ND_KEYS
)
}
\description{
Converts between weight and length measurements for a given species (using non-deterministic conversion keys, assuming these are known)
}
