% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEFAULT_IOTC_LL_EQUATIONS}
\alias{DEFAULT_IOTC_LL_EQUATIONS}
\title{DEFAULT_IOTC_LL_EQUATIONS}
\format{
A data frame
\describe{
  \item{\code{Time (s)}}{double COLUMN_DESCRIPTION}
  \item{\code{Acceleration - x (m/s²)}}{double COLUMN_DESCRIPTION}
  \item{\code{Acceleration - y (m/s²)}}{double COLUMN_DESCRIPTION}
  \item{\code{Acceleration - z (m/s²)}}{double COLUMN_DESCRIPTION}
  \item{\code{Acceleration - resultant (m/s²)}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
DEFAULT_IOTC_LL_EQUATIONS
}
\description{
DEFAULT_IOTC_LL_EQUATIONS
}
\keyword{datasets}
