% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEFAULT_IOTC_WL_ND_KEYS}
\alias{DEFAULT_IOTC_WL_ND_KEYS}
\title{DEFAULT_IOTC_WL_ND_KEYS}
\format{
A data frame
 \describe{
  \item{\code{SPECIES}}{character The IOTC species code}
  \item{\code{SOURCE_MEASURE_TYPE_CODE}}{character The IOTC weight measure code for the species' standard weight (one among: RND, GGT)}
  \item{\code{SOURCE_CLASS_LOW}}{integer the lower bound of the source weight class (in kg)}
  \item{\code{TARGET_MEASURE_TYPE_CODE}}{character The IOTC length measure code for the species' standard weight (one among: FL)}
  \item{\code{TARGET_CLASS_LOW}}{integer the lower bound of the target length class (in cm)}
  \item{\code{PROPORTION}}{double The proportion of fish allocated to the target weight class (converted from a source length class)}
}
}
\source{
\url{https://iotc.org/WPTT/24/Data/13-Equations/}
}
\usage{
DEFAULT_IOTC_WL_ND_KEYS
}
\description{
Keys to implement a non-deterministic weight-length conversion equations for BET and YFT
}
\keyword{datasets}
